Add types directory

+++ pages/api/update/index.ts
@@ -1,6 +1,18 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import prisma from '@/lib/prisma';
 
+interface UpdatePostRequestBody {
+  id: number;
+  title: string;
+  content: string;
+  slug: string;
+  teaser: string;
+  category: string;
+  featured: boolean;
+  editFeatured: boolean;
+  showEdited: boolean;
+}
+
 // POST /api/update
 export default async function handle(
   req: NextApiRequest,
@@ -16,13 +28,18 @@ export default async function handle(
     featured,
     editFeatured,
     showEdited,
-  } = req.body;
+  }: UpdatePostRequestBody = req.body;
 
   const post = await prisma.post.findUnique({
     where: {
       id: id,
     },
   });
+
+  if (!post) {
+    res.status(404).json({ message: `Post with ID ${id} not found` });
+    return;
+  }
 
   const postHistoryCount = await prisma.postHistory.count({
     where: { postId: post.id },
@@ -34,9 +51,11 @@ export default async function handle(
       orderBy: { editedAt: 'asc' },
     });
 
-    await prisma.postHistory.delete({
-      where: { id: oldestPostHistory.id },
-    });
+    if (oldestPostHistory) {
+      await prisma.postHistory.delete({
+        where: { id: oldestPostHistory.id },
+      });
+    }
   }
 
   if (showEdited) {
